% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{anexppop}
\alias{anexppop}
\title{Exponential population growth}
\usage{
anexppop(N0, b, d, t)
}
\arguments{
\item{N0}{Initial population size}

\item{b}{per capita birth rate}

\item{d}{per capita death rate}

\item{t}{time index, currently ignored.}
}
\value{
a single numeric value which is the next population size
}
\description{
A simple test function for the package. Deterministic exponential growth
with independent parameters for births and deaths. Only projects one time step.
}
