% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaLandWrappers.R
\name{createLandscape}
\alias{createLandscape}
\alias{plot.landscape}
\title{Constructor function for landscape class objects}
\usage{
createLandscape(parms = NULL)

\method{plot}{landscape}(rl, ...)
}
\arguments{
\item{parms}{list of arguments}
}
\value{
Object of class landscape
}
\description{
Constructor function for landscape class objects
}
\examples{
# calling the function without arguments gives a default landscape with 5 patches
rl <- createLandscape()
plot(rl)
}
