% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{exppop}
\alias{exppop}
\title{Exponential population growth}
\usage{
exppop(N0, parms)
}
\arguments{
\item{N0}{Initial population size}

\item{parms}{one row data_frame with model parameters}
}
\value{
a single numeric value which is the next population size
}
\description{
Exponential population growth
}
