% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{hh_A}
\alias{hh_A}
\alias{hh_N}
\title{Helmeted Honeyeater projection matrices}
\format{two lists
\describe{
  \item{hh_A}{ a list of 3 10x10 projection matrices.}
  \item{hh_N}{ a list of 4 vectors of length 10 with counts by age.}
}}
\source{
Unfortunately I have no idea where this data came from.
}
\usage{
hh_A

hh_N
}
\description{
3 age structured projection matrices for Helmeted Honeyeaters in 1991-1994, and
the associated population vectors.
}
\details{
This bird is critically endangered, restricted to 3 small populations
just east of Melbourne. Zoos Victoria contributes to conservation by captive
rearing individuals to supplement wild populations, as well as maintaining a
captive population "lifeboat" for the species.
}
\keyword{datasets}
