% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{iterate}
\alias{iterate}
\title{Iterate a population model}
\usage{
iterate(parms = NULL, N0 = NULL, popfun = NULL)
}
\arguments{
\item{parms}{a tibble containing the parameters of the model. Must have
at least column `t`, and one row for each time step.}

\item{N0}{a numeric vector of the initial population size. If named, these names
will be used in the returned tibble.}

\item{popfun}{a function that steps the model by one time step.
Must take at least N0 as an argument.}
}
\value{
a data_frame containing the input parameters and a column of projected population size
}
\description{
Iterate a population model forward in time from an initial population vector.
Parameters of the model can vary with time.
}
\examples{

}
