% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{iterate_pop}
\alias{iterate_pop}
\title{Iterate a population model}
\usage{
iterate_pop(parms = NULL, popfun = NULL)
}
\arguments{
\item{parms}{a data_frame containing the parameters of the model. Must have
arguments t and N0.}

\item{popfun}{a function that steps the model by one}
}
\value{
a data_frame containing the input parameters and a column of projected population size
}
\description{
Iterate a population model
}
\examples{

}
