% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypop-package.R
\docType{package}
\name{tidypop-package}
\alias{tidypop}
\alias{tidypop-package}
\title{tidypop: Tidy Interface to Single Species Population Models}
\description{
Facilitate teaching simple single species population models in a tidy context, and provide a 'tidy' interface based on data frames to existing population packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atyre2/tidypop}
  \item Report bugs at \url{https://github.com/atyre2/tidypop/issues}
}

}
\author{
\strong{Maintainer}: Drew Tyre \email{atyre2@unl.edu}

}
\keyword{internal}
