% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{iterate}
\alias{iterate}
\alias{iterate.tbl_df}
\alias{iterate.data.frame}
\alias{iterate.NULL}
\alias{iterate.default}
\title{Iterate a population model}
\usage{
iterate(...)

\method{iterate}{tbl_df}(parms = NULL, N0 = NULL, popfun = NULL)

\method{iterate}{data.frame}(parms = NULL, N0 = NULL, popfun = NULL)

\method{iterate}{NULL}(...)

\method{iterate}{default}(...)
}
\arguments{
\item{parms}{a data frame containing the parameters of the model. Must have
one row for each time step.}

\item{N0}{a numeric vector of the initial population size. If named, these names
will be used in the returned tibble.}

\item{popfun}{a function that steps the model by one time step.
Must take at least N0 as an argument.}
}
\value{
A data frame containing the input parameters and a column of projected population size.
The return value will have the same class as parms.
}
\description{
Iterate a population model forward in time from an initial population vector.
Parameters of the model can vary with time.
}
\details{
Will fail if any of the three input arguments is NULL. If `parms` only has
a single row, just column binds the initial population size. If `parms` has
zero rows, the result also has zero rows.
}
\examples{
# a simple exponential growth model
anexppop <- function(N0, b, d) {
  N1 <- N0 * (1 + b - d)
  return(N1)
}
# make the input dataframe
inputs <- data.frame(Year = 2018:2025, b = 0.2, d = 0.15)
iterate(inputs, 23, anexppop)
# time dependent model
inputs <- data.frame(Year = 2018:2025, b = seq(0.2, 0.1, length = 8), d = 0.15)
iterate(inputs, 23, anexppop)
}
